requirements.txt
bcrypt==4.0.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pyee==8.2.2
requests==2.28.1
soupsieve==2.3.2.post1
tqdm==4.64.1
urllib3==1.26.12
websockets==10.4
Werkzeug==2.2.2
zipp==3.10.0

python-dotenv~=0.21.0